requirements.txt
requests==2.7.0
twitter==1.17.1
ws4py==0.3.4
python-dateutil==2.4.2
six==1.9.0
